requirements.txt
psycopg2
Django
django-phonenumber-field
phonenumbers
django-heroku
gunicorn